+++ R/ggpar.R
@@ -12,8 +12,8 @@ NULL
 #'   "rickandmorty".
 #' @param gradient.cols vector of colors to use for n-colour gradient. Allowed
 #'   values include brewer and ggsci color palettes.
-#' @param main plot main title.
-#' @param submain plot subtitle.
+#' @param main,title plot main title.
+#' @param submain,subtitle plot subtitle.
 #' @param caption plot caption.
 #' @param xlab character vector specifying x axis labels, respectively. Use xlab
 #'   = FALSE to hide xlab.
@@ -25,6 +25,7 @@ NULL
 #'   respectively. For example \emph{font.x = c(14, "bold", "red")}. Use font.x
 #'   = 14, to change only font size; or use font.x = "bold", to change only font
 #'   face.
+#' @param font.title,font.subtitle alias of font.submain and font.submain, respectively.
 #' @param xlim,ylim a numeric vector of length 2, specifying  x and y axis
 #'   limits (minimum and maximum), respectively. e.g.: ylim = c(0, 50).
 #' @param xscale,yscale x and y axis scale, respectively. Allowed values are one
@@ -134,7 +135,9 @@ NULL
 #' @export
 ggpar <- function(p, palette = NULL, gradient.cols = NULL,
                   main = NULL, submain = NULL, caption = NULL, xlab = NULL, ylab = NULL,
+                  title = NULL, subtitle = NULL,
                   font.main = NULL, font.submain = NULL, font.x = NULL, font.y = NULL, font.caption = NULL,
+                  font.title = NULL, font.subtitle = NULL,
                   xlim = NULL, ylim = NULL,
                   xscale = c("none", "log2", "log10", "sqrt"),
                   yscale = c("none", "log2", "log10", "sqrt"),
@@ -154,6 +157,10 @@ ggpar <- function(p, palette = NULL, gradient.cols = NULL,
   if(is.ggplot(original.p)) list.plots <- list(original.p)
   else if(is.list(original.p)) list.plots <- original.p
   else stop("Can't handle an object of class ", class (original.p))
+  if(!is.null(title)) main <- title
+  if(!is.null(subtitle)) submain <- subtitle
+  if(!is.null(font.title)) font.main <- font.title
+  if(!is.null(font.subtitle)) font.submain <- font.subtitle
 
   for(i in 1:length(list.plots)){
     p <- list.plots[[i]]